+++ server/csv-renderers/statistics-flow-roles.js
@@ -18,11 +18,12 @@ module.exports = function (result, config) {
 		).join(',')
 	];
 
-	Object.keys(result).forEach(function (date) {
-		var resultItem = [mode.getDisplayedKey(new Date(date))];
+	Object.keys(result).forEach(function (rowName) {
+		var parsedDate = Date.parse(rowName);
+		var resultItem = isNaN(parsedDate) ? [rowName] : [mode.getDisplayedKey(new Date(parsedDate))];
 		Array.prototype.push.apply(resultItem,
-			Object.keys(result[date]).map(function (step) {
-				return (result[date][step] == null ? _("N/A") : result[date][step]);
+			Object.keys(result[rowName]).map(function (step) {
+				return (result[rowName][step] == null ? _("N/A") : result[rowName][step]);
 			}));
 		data.push(resultItem.join(','));
 	});
