+++ view/user-base.js
@@ -100,10 +100,11 @@ exports._submittedMenu = function () {
 	}
 	if (user.currentRoleResolved === 'statistics') {
 		return [li({ id: 'dashboard-nav' }, a({ href: '/' }, _("Dashboard"))),
-			li({ id: 'files-nav' }, a({ href: '/files/' }, _("Files"))),
+			li({ id: 'files-nav' }, a({ href: '/files/by-role/' }, _("Files"))),
 			li({ id: 'time-nav' }, a({ href: '/time/' }, _("Times"))),
 			li({ id: 'rejections-nav' }, a({ href: '/rejections/' }, _("Reasons of rejections"))),
-			li({ id: 'accounts-nav' }, a({ href: '/accounts/' }, _("Accounts")))];
+			li({ id: 'accounts-nav' }, a({ href: '/accounts/' }, _("Accounts"))),
+			li({ id: 'old-nav' }, a({ href: '/old/' }, _("Old")))];
 	}
 
 	isOfficialRole = user.officialRoles.has(user.currentRoleResolved);
